requirements.txt
python-dotenv
pandas
googlesearch-python
requests
beautifulsoup4
openai
streamlit
plotly
pinecone-client
tiktoken
